backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
scikit-learn>=1.5.0
numpy>=2.1.0
pillow>=10.4.0
pyyaml==6.0.1
requests==2.31.0
kaggle==1.6.6
gdown==4.7.1

# (opcional) pandas não é necessário para o treino do classificador
# pandas==2.2.3

# NOVAS DEPENDÊNCIAS PARA OTIMIZAÇÕES
albumentations==1.3.1
matplotlib>=3.9.0
seaborn==0.13.0
tqdm==4.66.1
psutil==5.9.6
opencv-python-headless==4.12.0.88

# Ortomosaico e tiling
stitching>=0.6.0
opencv-contrib-python>=4.8.0
sahi>=0.11.0
